confirm #32 soohyung

+++ src/main/webapp/WEB-INF/views/admin/bookDetail.jsp
@@ -10,16 +10,30 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <script>
-	function deleteCheck(id) {
-		var page = ${currentPage};
-		alert("정말 삭제하시겠습니까?");
+function updateCheck(){
+	if(confirm("변경 사항을 저장하시겠습니까?")){
+		document.bookForm.submit();
+	}else{
+		return false;
+	}
+}
+function deleteCheck(id) {
+	var page = ${currentPage};
+	if(confirm("정말 삭제하시겠습니까?")){
 		document.location.href="<%=cp%>/admin/bookDelete.do?book_num="+id+"&currentPage="+page;
+	}else{
+		return false;
 	}
-	function deleteCheck2(id, num) {
-		var page = ${reviewPage};
-		alert("정말 삭제하시겠습니까?");
+}
+function deleteCheck2(id) {
+	var page = ${reviewPage};
+	var num = ${view.book_num}
+	if(confirm("정말 삭제하시겠습니까?")){
 		document.location.href="<%=cp%>/admin/reviewDelete.do?review_num="+id+"&book_num="+num+"&currentPage="+page;
+	}else{
+		return false;
 	}
+}
 </script>
 </head>
 <body>
@@ -32,7 +46,14 @@
 					<li class="breadcrumb-item"><a href="<%=cp%>/admin/main.do">Home</a></li>
 					<li class="breadcrumb-item"><a
 						href="<%=cp%>/admin/bookList.do">Book List</a></li>
-					<li class="breadcrumb-item active">Book Detail</li>
+					<c:choose>
+						<c:when test="${view.book_num==0}">
+							<li class="breadcrumb-item active">Book Write</li>
+						</c:when>
+						<c:otherwise>
+							<li class="breadcrumb-item active">Book Detail</li>
+						</c:otherwise>
+					</c:choose>
 				</ol>
 			</div>
 		</div>
@@ -41,10 +62,19 @@
 				<h3>Book Information</h3>
 				<div class="card">
 					<div class="card-block">
-						<form class="form-horizontal form-material">
+						<form class="form-horizontal form-material" name="bookForm" method="POST" onsubmit="return updateCheck()" enctype="multipart/form-data">
 							<div class="row">
 								<div class="col-lg-4" style="margin: auto; text-align: center;">
 									<img class="img-fluid" src="<%=cp%>/upload/${view.book_image}" />
+									<input type="file" class="fileinput" name="book_image"/>
+									<c:choose>
+										<c:when test="${view.book_num==0}">
+											<input type="hidden" name="book_image" value="">
+										</c:when>
+										<c:otherwise>
+											<input type="hidden" name="book_image" value="${view.book_image}">
+										</c:otherwise>
+									</c:choose>
 								</div>
 								<div class="col-lg-8">
 									<div class="form-group"></div>
@@ -63,7 +93,7 @@
 									<div class="form-group">
 										<label class="col-md-12">Book Name</label>
 										<div class="col-md-12">
-											<input type="text" class="form-control form-control-line"
+											<input type="text" name="book_name" class="form-control form-control-line"
 												value="${view.book_name}">
 										</div>
 									</div>
@@ -72,14 +102,14 @@
 											<div class="col-md-6">
 												<label class="col-md-6">Category</label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_category}"
+													<input type="text" name="book_category" value="${view.book_category}"
 														class="form-control form-control-line">
 												</div>
 											</div>
 											<div class="col-md-6">
 												<label class="col-md-6">Author</label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_auth}"
+													<input type="text" name="book_auth" value="${view.book_auth}"
 														class="form-control form-control-line">
 												</div>
 											</div>
@@ -90,14 +120,14 @@
 											<div class="col-md-6">
 												<label class="col-md-6">Company</label>
 												<div class="col-md-12">
-													<input type="text" value="${view.company_id}"
+													<input type="text" name="company_id" value="${view.company_id}"
 														class="form-control form-control-line">
 												</div>
 											</div>
 											<div class="col-md-6">
 												<label class="col-md-6">Publish Date</label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_publish_date}"
+													<input type="text" name="book_publish_date" value="${view.book_publish_date}"
 														class="form-control form-control-line">
 												</div>
 											</div>
@@ -108,28 +138,28 @@
 											<div class="col-md-3">
 												<label class="col-md-12">Price</label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_price}"
+													<input type="text" name="book_price" value="${view.book_price}"
 														class="form-control form-control-line">
 												</div>
 											</div>
 											<div class="col-md-3">
 												<label class="col-md-12"><b>판매량</b></label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_sell_count}"
+													<input type="text" name="book_sell_count" value="${view.book_sell_count}"
 														class="form-control form-control-line">
 												</div>
 											</div>
 											<div class="col-md-3">
 												<label class="col-md-12"><b>현재 재고</b></label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_current_count}"
+													<input type="text" name="book_current_count" value="${view.book_current_count}"
 														class="form-control form-control-line">
 												</div>
 											</div>
 											<div class="col-md-3">
 												<label class="col-md-12"><b>기초 재고</b></label>
 												<div class="col-md-12">
-													<input type="text" value="${view.book_base_count}"
+													<input type="text" name="book_base_count" value="${view.book_base_count}"
 														class="form-control form-control-line">
 												</div>
 											</div>
@@ -141,14 +171,13 @@
 								<label class="col-md-12">Content</label>
 								<div class="form-group"></div>
 								<div class="col-md-12">
-									<textarea rows="20" class="form-control form-control-line">${view.book_content}</textarea>
+									<textarea rows="20" name="book_content" class="form-control form-control-line">${view.book_content}</textarea>
 								</div>
 							</div>
 							<div class="form-group">
 								<div class="col-sm-12" style="text-align: center;">
 									<div class="row">
-										<label class="col-sm-4"><b>활성화
-												여부</b></label>
+										<label class="col-sm-4"><b>활성화 여부</b></label>
 										<div class="col-sm-8">
 											<input type="radio" name="book_use_yn" id="book_use_yn"
 												value="1"
@@ -164,12 +193,17 @@
 							<div class="form-group"></div>
 							<div class="form-group">
 								<div class="col-sm-12" style="text-align: center;">
-									<button class="btn btn-warning"
-										onclick="javascript:location.href='<%=cp%>/admin/bookModify.do?book_num=${view.book_num}&currentPage=${currentPage}'">Update
-										Profile</button>&nbsp;
-									<button class="btn btn-danger"
-										onclick="deleteCheck(${view.book_num})">Delete
-										Profile</button>
+									<c:choose>
+										<c:when test="${view.book_num==0}">
+											<button class="btn btn-success">Write</button>
+										</c:when>
+										<c:otherwise>
+											<button class="btn btn-warning">Update</button>
+											&nbsp;
+											<a class="btn btn-danger"
+												href="javascript:deleteCheck(${view.book_num})">Delete</a>
+										</c:otherwise>
+									</c:choose>
 								</div>
 							</div>
 						</form>
@@ -177,6 +211,7 @@
 				</div>
 			</div>
 		</div>
+		<c:if test="${view.book_num!=0}">
 		<div class="col-lg-12">
 			<h3>Book Review</h3>
 			<div class="card">
@@ -194,18 +229,17 @@
 										<c:url var="viewURL" value="/admin/memberDetail.do">
 											<c:param name="member_id" value="${list.member_id}" />
 										</c:url>
-											<tr>
-												<td>${list.review_num}</td>
-												<td>
-													<!-- member_name review에서 조인하기.. --> (<a href="${viewURL}">${list.member_id}</a>)
-												</td>
-												<td>${list.review_pw}</td>
-												<td>${list.review_regdate}</td>
-												<td rowspan="2">
-													<a class="mdi mdi-delete" title="Delete" onclick="deleteCheck2(${list.review_num}, ${view.book_num})"></a>
-												</td>
-											</tr>
-											
+										<tr>
+											<td>${list.review_num}</td>
+											<td>
+												<!-- member_name review에서 조인하기.. --> (<a href="${viewURL}">${list.member_id}</a>)
+											</td>
+											<td>${list.review_pw}</td>
+											<td>${list.review_regdate}</td>
+											<td rowspan="2"><a class="mdi mdi-delete" title="Delete"
+												href="javascript:deleteCheck2(${list.review_num})"></a></td>
+										</tr>
+
 										<tr>
 											<td colspan=3 style="width: 400px;">${list.review_content}</td>
 											<td>${list.star_point}</td>
@@ -224,6 +258,7 @@
 				</div>
 			</div>
 		</div>
+		</c:if>
 	</div>
 	<!-- ============================================================== -->
 	<!-- footer -->
